% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slobe-package.R
\docType{package}
\name{slobe-package}
\alias{slobe-package}
\title{slobe: SLOBE models}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Joanna Pokora \email{pokorajo@gmail.com}

Authors:
\itemize{
  \item Joanna Pokora \email{pokorajo@gmail.com}
}

}
\keyword{internal}
