% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slobe.R
\name{slobe}
\alias{slobe}
\title{SLOBE}
\usage{
slobe(
  X,
  y,
  family = "gaussian",
  start = NULL,
  a_prior = 0.01 * NROW(X),
  b_prior = 0.01 * NROW(X),
  Covmat = NULL,
  sigma = NULL,
  FDR = 0.05,
  tol = 1e-04,
  max_iter = 100L,
  verbose = FALSE,
  BH = TRUE
)
}
\arguments{
\item{X}{design matrix}

\item{y}{numeric, response variable}

\item{family}{model family - either "gaussian" (default) or "binomial"}

\item{start}{model coefficients used for initialization}

\item{a_prior, b_prior}{non-negative parameters of the prior Beta distribution
on theta}

\item{Covmat}{numeric covariance matrix, default to identity matrix}

\item{sigma}{the variance of the noise, default to 1}

\item{FDR}{False Discovery Rate, default to 0.05}

\item{tol}{optimization tolerance}

\item{max_iter}{the maximal number of iterations of the optimization
algorithm, default to 100}

\item{verbose}{verbosity, default to FALSE}

\item{BH}{logical, indicates whether the Benjamini-Hochberg correction for
multiple testing should be used}
}
\description{
Fits a Gaussian or logistic model regularized with
Adaptive Bayesian SLOPE
}
\examples{
set.seed(17)
xy <- slobe:::generate_data(signals_num = 20, signals_strength = 5)
fit <- slobe(xy$X, xy$y)
fit$coefficients
}
\references{
Jiang, W., Bogdan, M., Josse, J., Majewski, S., Miasojedow, B.,
Ročková, V., & TraumaBase® Group. (2021). Adaptive Bayesian SLOPE: Model
Selection with Incomplete Data. Journal of Computational and Graphical
Statistics, 1-25. \doi{10.1080/10618600.2021.1963263}

Bogdan, M., van den Berg, E., Sabatti, C., Su, W., & Candès, E. J. (2015).
SLOPE -- adaptive variable selection via convex optimization. The Annals of
Applied Statistics, 9(3), 1103–1140. \doi{10/gfgwzt}

Ročková, V., & George, E. I. (2018). The spike-and-slab lasso. Journal of
the American Statistical Association, 113(521), 431-444.
\doi{10.1080/01621459.2016.1260469}
}
